немного изменен PolygonServiceImpl: теперь учтены 10% и 35% от спроса в случае жилой и коммерческой застройки соответственно

+++ src/main/java/ru/mephi/backend/dto/Coordinate.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Coordinate {
-    private double latitude;  // Широта
-    private double longitude; // Долгота
+    private float latitude;  // Широта
+    private float longitude; // Долгота
 }
